% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lplot.R
\name{lplot}
\alias{lplot}
\title{Creates plot of lift curve}
\usage{
lplot(data, predictions, actuals, numbins)
}
\arguments{
\item{data}{the dataset}

\item{predictions}{the predicted values}

\item{actuals}{the actual values}

\item{numbins}{the number of bins for the lift curve}
}
\value{
a plot
}
\description{
Creates plot of lift curve
}
\examples{
mock_data <- data.frame(preds = rep(1:4,10),target=rep(1:5,8),weights=rep(1:2,20))
lplot(mock_data,"preds","target",3)

}
