% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltable.R
\name{ltable}
\alias{ltable}
\title{Creates table for unweighted lift curve}
\usage{
ltable(data, predictions, actuals, numbins)
}
\arguments{
\item{data}{the dataset}

\item{predictions}{the predicted values}

\item{actuals}{the actual values}

\item{numbins}{the number of bins for the lift curve}
}
\value{
a table
}
\description{
Creates table for unweighted lift curve
}
\examples{
mock_data <- data.frame(preds = rep(1:4,10),target=rep(1:5,8),weights=rep(1:2,20))
ltable(data=mock_data,predictions="preds",actuals="target",numbins=3)

}
