% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlcredplot.R
\name{wlcredplot}
\alias{wlcredplot}
\title{Creates plot of weighted lift curve with credibility}
\usage{
wlcredplot(data, predictions, actuals, weights, numbins)
}
\arguments{
\item{data}{the dataset}

\item{predictions}{the predicted values}

\item{actuals}{the actual values}

\item{weights}{the weights for the values}

\item{numbins}{the number of bins for the lift curve}
}
\value{
a plot
}
\description{
Creates plot of weighted lift curve with credibility
}
\examples{
mock_data <- data.frame(preds = rep(1:4,10),target=rep(1:5,8),weights=rep(1:2,20))
wlcredplot(mock_data,"preds","target","weights",3)

}
