% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlplot.R
\name{wlplot}
\alias{wlplot}
\title{Creates plot of weighted lift curve}
\usage{
wlplot(data, predictions, actuals, weights, numbuckets)
}
\arguments{
\item{data}{the dataset}

\item{predictions}{the predicted values}

\item{actuals}{the actual values}

\item{weights}{the weights for the values}

\item{numbuckets}{the number of buckets for the lift curve}
}
\value{
a plot
}
\description{
Creates plot of weighted lift curve
}
\examples{
mock_data <- data.frame(preds = rep(1:4,10),target=rep(1:5,8),weights=rep(1:2,20))
wlplot(mock_data,"preds","target","weights",3)

}
