% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wltable.R
\name{wltable}
\alias{wltable}
\title{Creates table for weighted lift curve}
\usage{
wltable(data, predictions, actuals, weights, numbuckets)
}
\arguments{
\item{data}{the dataset}

\item{predictions}{the predicted values}

\item{actuals}{the actual values}

\item{weights}{the weights for the values}

\item{numbuckets}{the number of buckets for the lift curve}
}
\value{
a table
}
\description{
Creates table for weighted lift curve
}
\examples{
mock_data <- data.frame(preds = rep(1:4,10),target=rep(1:5,8),weights=rep(1:2,20))
wltable(mock_data,"preds","target","weights",3)

}
